requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==0.7.15
alembic==1.13.1
annotated-types==0.6.0
any==0.0.1
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
astroid==3.2.4
asttokens==2.4.1
async-lru==2.0.4
asyncssh==2.17.0
atomicwrites==1.4.1
attrs==23.2.0
autopep8==2.0.4
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
bidict==0.23.1
binaryornot==0.4.4
black==23.12.1
bleach==6.1.0
blinker==1.9.0
buffer==0.2.1
build==1.2.1
cachetools==5.5.2
camelot-py==1.0.9
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cookiecutter==2.5.0
crispy-tailwind==0.5.0
cryptography==41.0.7
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
diff-match-patch==20230430
dill==0.3.8
distro==1.9.0
Django==5.0.1
django-crispy-forms==2.0
django-filter==23.2
django-mysql==4.14.0
djangorestframework==3.14.0
docstring-to-markdown==0.13
docutils==0.20.1
docx2pdf==0.1.8
easyocr==1.7.2
et-xmlfile==1.1.0
excel==1.0.1
executing==2.0.1
fastapi==0.110.3
fastjsonschema==2.19.1
filelock==3.20.0
flake8==7.1.1
Flask==3.1.1
fonttools==4.47.0
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2025.12.0
ghostscript==0.7
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.26.0
idna==3.6
ImageIO==2.37.2
imagesize==1.4.1
img2pdf==0.6.3
importlib-metadata==7.0.1
inflection==0.5.1
instagram==1.3.4
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.3.0
jedi==0.19.1
jellyfish==1.0.3
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keyring==24.3.0
kiwisolver==1.4.5
lab==8.5
lazy_loader==0.4
lxml==5.3.0
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.0.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.4.0
mysqlclient==2.2.1
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.6
nh3==0.2.17
ninja==1.11.1.4
notebook==7.0.6
notebook_shim==0.2.3
numpy==2.2.6
numpydoc==1.6.0
oauthlib==3.2.2
ocrmypdf==16.12.0
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
openpyxl==3.1.2
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
panda==0.3.1
pandas==2.2.2
pandocfilters==1.5.0
paramiko==3.4.0
parso==0.8.3
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20251107
pdfplumber==0.11.8
pexpect==4.9.0
pi_heif==1.1.1
pickleshare==0.7.5
pikepdf==10.0.2
pillow==12.0.0
pipdeptree==2.16.2
pkginfo==1.10.0
platformdirs==4.1.0
pluggy==1.3.0
prometheus-client==0.19.0
prompt==0.4.1
prompt-toolkit==3.0.43
proto-plus==1.26.0
protobuf==5.29.3
psutil==5.9.7
psycopg2==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pyclipper==1.4.0
pycodestyle==2.12.1
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pydocstyle==6.3.0
pyflakes==3.2.0
PyGetWindow==0.0.9
PyGithub==2.4.0
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.2.7
pylint-venv==3.0.3
pyls-spyder==0.4.0
PyMsgBox==1.0.9
PyMuPDF==1.26.6
PyMySQL==1.1.1
PyNaCl==1.5.0
pyodbc==5.1.0
pyparsing==3.1.1
pypdf==5.7.0
PyPDF2==3.0.1
pypdfium2==5.1.0
pyperclip==1.9.0
pyproject_hooks==1.1.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.8.2
python-docx==1.2.0
python-dotenv==1.1.1
python-engineio==4.9.0
python-json-logger==2.0.7
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
python-multipart==0.0.9
python-slugify==8.0.1
python-socketio==5.11.2
pytoolconfig==1.2.6
pytweening==1.2.0
pytz==2023.3.post1
pyuca==1.2
pywhatkit==5.4
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
QDarkStyle==3.2.3
qstylizer==0.2.2
QtAwesome==1.3.1
qtconsole==5.6.1
QtPy==2.4.1
readme_renderer==43.0
redis==5.0.4
referencing==0.32.1
reflex==0.4.9
reflex-hosting-cli==0.1.13
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.0
rope==1.11.0
rpds-py==0.16.2
rsa==4.9
Rtree==1.1.0
scapy==2.6.1
scikit-image==0.25.2
scipy==1.16.3
selenium==4.29.0
Send2Trash==1.8.2
setuptools==69.5.1
shapely==2.1.2
shellingham==1.5.4
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
spyder-kernels==3.0.3
sql==2022.4.0
SQLAlchemy==2.0.29
sqlmodel==0.0.18
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
starlette-admin==0.13.2
statistics==1.0.3.5
superqt==0.6.7
sympy==1.14.0
tabula-py==2.10.0
tabulate==0.9.0
tailwind==3.1.5b0
telegram==0.0.1
terminado==0.18.0
text-unidecode==1.3
textdistance==4.6.1
three-merge==0.1.1
tifffile==2025.10.16
tinycss2==1.2.1
tkinterdnd2==0.4.3
tomli==2.0.1
tomlkit==0.12.4
torch==2.9.1
torchvision==0.24.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
trio==0.29.0
trio-websocket==0.12.2
tube==1.2.1
twilio==9.0.0
twine==5.0.0
txt2tags==3.9
typer==0.12.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.11.0
tzdata==2023.4
ujson==5.9.0
unicode==2.9
Unidecode==1.3.8
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
whatthepatch==1.0.5
wheel==0.43.0
widgetsnbextension==4.0.9
wikipedia==1.4.0
wrapt==1.11.2
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.1.9
xmltodict==0.13.0
yapf==0.40.2
yarl==1.9.4
you==0.5.0.dev1
zipp==3.17.0
